pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "bio_extraction"
version = "0.1.0"
description = "Biographical entity extraction pipeline for Polish historical PDFs"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "requests>=2.31",
    "pymupdf>=1.27.2.2",
    "mmh3>=5.2.1",
    "warcio>=1.7.5",
    "bitarray>=3.8.0",
    "numpy>=2.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.5",
    "pre-commit>=3.7",
    "types-PyYAML>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bio_extraction*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest-mock>=3.15.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20260107",
]
